all function return integers except divide

+++ arithmetic.py
@@ -4,19 +4,19 @@
 def add(num1, num2):
     """Return the sum of the two input integers."""
 
-    return num1 + num2
+    return int(num1 + num2)
 
 
 def subtract(num1, num2):
     """Return the second number subtracted from the first."""
 
-    return num1 - num2
+    return int(num1 - num2)
 
 
 def multiply(num1, num2):
     """Multiply the two inputs together."""
 
-    return num1 * num2
+    return int(num1 * num2)
 
 
 def divide(num1, num2):
@@ -28,22 +28,22 @@ def divide(num1, num2):
 def square(num1):
     """Return the square of the input."""
 
-    return num1 ** 2
+    return int(num1 ** 2)
 
 
 def cube(num1):
     """Return the cube of the input."""
 
-    return num1 ** 3
+    return int(num1 ** 3)
 
 
 def power(num1, num2):
     """Raise num1 to the power of num2 and return the value."""
 
-    return num1 ** num2
+    return int(num1 ** num2)
 
 
 def mod(num1, num2):
     """Return the remainder of num1 / num2."""
 
-    return num1 % num2
+    return int(num1 % num2)
